requirements.txt
# Core dependencies
pyyaml>=6.0
pydantic>=2.0

# Web framework
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
anyio>=4.0.0

# Development
python-multipart>=0.0.6
